Factor out log collection
